xlsx and xls to txt update
Updated the script that translates xlsx and xls submitted files to txt files for UT to process. This update will replace the values found in the excel files to a <> value where the source configuration has set to NOT TO USE field. Some sites submit paragraph values in the excel fields for field we are not using, this was breaking UT, so becaue the fields are marked as do not use we replace the incoming value to blank so we can process.